Complete rehaul of zooming logic.
Using the CATransform3D instead of CGAffineTransform -- this way
autolayout does not mess anything up (with or without it).

Also, rather than using the passed in container view's superview for
zooming logic, using the [UIScreen mainScreen].bounds so that the
picture will zoom to the full screen size when tapped.